+++ include/BuyTowerScreen.hpp
@@ -21,19 +21,22 @@ private:
   int row;
   //int col clicked on
   int col;
+  //whether we are buying or selling (only applies when there is a tower)
+  bool areBuying;
 
 public:
   BuyTowerScreen(shared_ptr<EventManager> eventManager,shared_ptr<TextLoader> textLoader, shared_ptr<GameLogic> gameLogic,int windowX, int windowY);
   void initText();
   void changeTitleString(string newTitle);
 
-  void setSelectedTile(int row, int col);
   void setTextToNewTile();
 
   void draw(sf::RenderWindow& window);
   void drawTitle(sf::RenderWindow& window);
   void registerDelegates();
+  void registerPersistentDelegates();
   void deregisterDelegates();
   void handleKeyPress(const EventInterface& event);
   void handleMousePress(const EventInterface& event);
+  void handleStateChange(const EventInterface& event);
 };