+++ include/Events/WaveChangeEventData.hpp
@@ -7,8 +7,8 @@ using std::string;
 
 class WaveChangeEventData : public EventDataInterface{
 public:
-  int waveID;
+  int waveNum;
   float timeStamp;
 
-  WaveChangeEventData (int waveID, float timeStamp);
+  WaveChangeEventData (int waveNum, float timeStamp);
 };