+++ src/BuyTowerScreen.cpp
@@ -6,8 +6,8 @@ BuyTowerScreen::BuyTowerScreen(shared_ptr<EventManager> eventManager,shared_ptr<
   this -> eventManager = eventManager;
   this -> textLoader = textLoader;
   this -> gameLogic = gameLogic;
-
   this -> initText();
+  this -> registerPersistentDelegates();
 }
 
 /*
@@ -25,23 +25,21 @@ void BuyTowerScreen::changeTitleString(string newTitle){
 }
 
 /*
- * Set the selected tile for this buyTowerScreen
- */
-void BuyTowerScreen::setSelectedTile(int row, int col){
-  this -> row = row;
-  this -> col = col;
-
-  setTextToNewTile();
-}
-
-/*
  * Set the current title text to reflect what is located at the
  * selected tile position
  */
  void BuyTowerScreen::setTextToNewTile(){
+   cout << row << " " << col << endl;
+
    //change what the buyTower button says depending on the what is at the tile
    if(gameLogic->isTower(row,col)){
-     changeTitleString(textLoader->getString("IDS_Buy_Tower_Title_Text_Existing_Tower_Or_Empty_Space"));
+     //if we are currently buying the tower then we set differnet text than if we are selling one
+     if(areBuying){
+          changeTitleString(textLoader->getString("IDS_Buy_Tower_Title_Text_Existing_Tower_Or_Empty_Space"));
+     }
+     else{
+          changeTitleString(textLoader->getString("IDS_Buy_Tower_Title_Text_Existing_Tower_To_Sell"));
+     }
    }
    else if(gameLogic->isObstacle(row,col)){
      changeTitleString(textLoader->getString("IDS_Buy_Tower_Existing_Obstacle"));
@@ -110,6 +108,24 @@ void BuyTowerScreen::registerDelegates(){
   this -> eventManager -> registerDelegate(mousePressDelegate, textLoader -> getString(string("IDS_BTS_MP")),mousePressEventType);
 
 }
+
+/*
+ * Register the delegate methods for this class
+ * with any events it needs to know about
+ * used for any delegate methods that should never be deregistered
+ */
+void BuyTowerScreen::registerPersistentDelegates(){
+
+  //bind our delegate function for mouse presses
+  EventManager::EventDelegate stateChangeDelegate = std::bind(&BuyTowerScreen::handleStateChange, this, _1);
+
+  //make an event and get its type
+  StateChangeEvent stateChangeEvent = StateChangeEvent();
+  EventType stateChangeEventType = stateChangeEvent.getEventType();
+  //register the delegate and its type
+  this -> eventManager -> registerDelegate(stateChangeDelegate, textLoader -> getString(string("IDS_BTD_SC")),stateChangeEventType);
+}
+
 /*
  * Deregisetr the delegated methods for this class
  * so they are not called when we switch off this screen
@@ -182,5 +198,42 @@ void BuyTowerScreen::handleMousePress(const EventInterface& event){
   float xPos = mpEventData -> x;
   //get the y position
   float yPos = mpEventData -> y;
+
+}
+
+/*
+ * Handle a state change
+ * @param event: event of the state change
+ */
+void BuyTowerScreen::handleStateChange(const EventInterface& event){
+  /*
+   * cast the EventInterface reference to a CONST pointer to the
+   * StateChangeEvent type which allows us to access variables and methods
+   * specific to StateChangeEvent
+   */
+  const StateChangeEvent* scEvent = static_cast<const StateChangeEvent*>(&event);
+  /*
+   * cast the "data" (a EventDataInterface) to a StateChangeEventData type
+   * the .get() is because data is a unique_ptr and we need to grab the
+   * raw pointer inside of it for this
+   */
+  StateChangeEventData* scEventData = static_cast<StateChangeEventData*>((scEvent -> data).get());
+
+  State state = scEventData -> state;
+
+  //we will only care if the state is the current one
+  //because there will be data in it that we need
+  if(state != State::BuyTower){
+    return;
+  }
+
+  row = scEventData -> row;
+  col = scEventData -> col;
+  areBuying = scEventData -> buy;
+
+  //update the text on the screen to reflect the circumstances of the row, col
+  //i.e. is this a tower we are upgrading or selling, is this an obstacle we are
+  //removing
+  setTextToNewTile();
 
 }