+++ src/Events/StateChangeEvent.cpp
@@ -29,10 +29,11 @@ StateChangeEvent::StateChangeEvent(State state, float timeStamp){
  * @param state: the enum for the state
  * @param row: the current row selected
  * @param col: the current col selected
+ * @param buy: whether we are buying a tower or upgrading or removing an obstacle (false when selling a tower)
  * @param timeStamp: the timestamp of the creation of this event
  */
-StateChangeEvent::StateChangeEvent(State state, int row, int col, float timeStamp){
-  this -> data = unique_ptr<StateChangeEventData>(new StateChangeEventData(state, row, col, timeStamp));
+StateChangeEvent::StateChangeEvent(State state, int row, int col, bool buy, float timeStamp){
+  this -> data = unique_ptr<StateChangeEventData>(new StateChangeEventData(state, row, col, buy, timeStamp));
 }
 
 /*