+++ src/Events/StateChangeEventData.cpp
@@ -10,9 +10,10 @@
    this -> timeStamp = timeStamp;
  }
 
- StateChangeEventData::StateChangeEventData(State state, int row, int col, float timeStamp){
+ StateChangeEventData::StateChangeEventData(State state, int row, int col, bool buy, float timeStamp){
    this ->state = state;
    this ->row = row;
    this ->col = col;
+   this ->buy = buy;
    this -> timeStamp = timeStamp;
  }