+++ src/PlayingScreenHeader.cpp
@@ -7,13 +7,13 @@ PlayingScreenHeader::PlayingScreenHeader(shared_ptr<EventManager> eventManager, 
   this->gameLogic=gameLogic;
   this -> windowX = gameLogic -> getWindowX();
   this-> windowY = gameLogic -> getWindowY();
-  string fontpath = textLoader -> getString(string("IDS_FFP"));
+  string fontpath = textLoader -> getString(string("IDS_TFP"));
   this -> buyTower = unique_ptr<Button>(new Button(windowX, windowY, TOPRIGHT,
     textLoader -> getString(string("IDS_Buy_Tower_Empty_Space")), textLoader, fontpath));
 
   visible = true;
   initDrawingMaterials();
-
+  recaculateHeader = false;
   ySize = getMaximumY();
 }
 
@@ -63,7 +63,7 @@ void PlayingScreenHeader::initBuyTowerButton(){
   //set the text character size
   this->buyTower -> setTextSize(this->windowX / this->textLoader->getInteger(string("IDS_Buy_Tower_Text_Size")));
 
-  string fontpath = textLoader -> getString(string("IDS_FFP"));
+  string fontpath = textLoader -> getString(string("IDS_TFP"));
   //set the font inside the button so it can be used to calculate a bounds
   (this->buyTower) -> setFont(fontpath);
 
@@ -78,9 +78,81 @@ void PlayingScreenHeader::initBuyTowerButton(){
  * Set up the sell tower button
  */
 void PlayingScreenHeader::initSellTowerButton(){
-  string fontpath = textLoader -> getString(string("IDS_FFP"));
-  this -> sellTower = unique_ptr<Button>(new Button(windowX, windowY, TOPRIGHT,
+  //padding between each button in header
+  int headerButtonPaddingX = textLoader -> getInteger(string("IDS_Playing_Screen_Header_Padding_X"));
+  int headerButtonPaddingY = textLoader -> getInteger(string("IDS_Playing_Screen_Header_Padding_Y"));
+
+  //the bounding rectangle for the rectangle around the balance button
+  sf::RectangleShape offSetRect = (this->buyTower) -> getButtonRect();
+
+  //position
+  sf::Vector2f pos = offSetRect.getPosition();
+  //dimensions
+  sf::Vector2f dimensions = offSetRect.getSize();
+
+  //the xPosition is the length of the balance button plus some padding
+  float xPos = pos.x;
+  //the yPosition is the
+  float yPos =  pos.y + dimensions.y + headerButtonPaddingY;
+
+  string fontpath = textLoader -> getString(string("IDS_TFP"));
+  this -> sellTower = unique_ptr<Button>(new Button(windowX, windowY, xPos, yPos,
     textLoader -> getString(string("IDS_Sell_Tower")), textLoader, fontpath));
+
+  // set the fill color for the button rectangle
+  this -> sellTower -> setFillColor(this->textLoader -> getInteger(string("IDS_Sell_Tower_Fill_Color_Red")),
+  this->textLoader -> getInteger(string("IDS_Sell_Tower_Fill_Color_Blue")), this->textLoader -> getInteger(string("IDS_Sell_Tower_Fill_Color_Green")),
+  this->textLoader -> getInteger(string("IDS_Sell_Tower_Fill_Color_Alpha")));
+
+  // set the outline color for the button
+  this -> sellTower -> setOutlineColor(this->textLoader -> getInteger(string("IDS_Sell_Tower_Outline_Color_Red")),
+  this->textLoader -> getInteger(string("IDS_Sell_Tower_Outline_Color_Blue")),this->textLoader -> getInteger(string("IDS_Sell_Tower_Outline_Color_Green")),
+  this->textLoader -> getInteger(string("IDS_Sell_Tower_Outline_Color_Alpha")));
+
+  //set the button outline thickness
+  //this->sellTower -> setOutlineThickness(this->textLoader -> getInteger(string("IDS_Sell_Tower_Outline_Thickness")));
+
+  //set the fill color for the button text
+  this -> sellTower -> setTextFillColor(this->textLoader -> getInteger(string("IDS_Sell_Tower_Text_Fill_Color_Red")),
+  this->textLoader -> getInteger(string("IDS_Sell_Tower_Text_Fill_Color_Blue")), this->textLoader -> getInteger(string("IDS_Sell_Tower_Text_Fill_Color_Green")),
+  this->textLoader -> getInteger(string("IDS_Sell_Tower_Text_Fill_Color_Alpha")));
+
+
+  //set the outline color for the text
+  this -> sellTower -> setTextOutlineColor(this->textLoader -> getInteger(string("IDS_Sell_Tower_Text_Outline_Color_Red")),
+  this->textLoader -> getInteger(string("IDS_Sell_Tower_Text_Outline_Color_Blue")),this->textLoader -> getInteger(string("IDS_Sell_Tower_Text_Outline_Color_Green")),
+  this->textLoader -> getInteger(string("IDS_Sell_Tower_Text_Outline_Color_Alpha")));
+
+  //set the button text outline thickness
+  this->sellTower -> setTextOutlineThickness(this->textLoader -> getInteger(string("IDS_Sell_Tower_Text_Outline_Thickness")));
+
+  //set the text character size
+  this->sellTower -> setTextSize(this->windowX / this->textLoader->getInteger(string("IDS_Sell_Tower_Text_Size")));
+
+  //set the font inside the button so it can be used to calculate a bounds
+  (this->sellTower) -> setFont(fontpath);
+
+
+  //was not setting itself properly
+  //the bounding rectangle for the rectangle around the balance button
+  sf::RectangleShape ownRect = (this->sellTower) -> getButtonRect();
+
+  //position
+  sf::Vector2f ownRectPos = ownRect.getPosition();
+  //dimensions
+  sf::Vector2f ownRecDimensions = ownRect.getSize();
+
+  //the xPosition is the length of the balance button plus some padding
+  float newxPos = pos.x;
+  //the yPosition is the
+  float newyPos =  dimensions.y + ownRecDimensions.y;
+
+  //rescale the button and reset it
+  (this -> sellTower) -> scaleButton( newxPos, newyPos);
+
+  //make the button invisible to begin with
+  (this->sellTower) -> flipVisibility();
+
 }
 
 /*
@@ -144,6 +216,9 @@ void PlayingScreenHeader::initHitpointsButton(){
 
   //push this button into the vector of buttons to be drawn
   headerVariables.push_back(hitpoints);
+
+  //push the initial value of the button
+  headerVariableValues.push_back(startingHitpoints);
 }
 
 /*
@@ -173,7 +248,7 @@ void PlayingScreenHeader::initBalanceButton(){
   float yPos =  headerButtonPaddingY;
 
 
-  string fontpath = textLoader -> getString(string("IDS_FFP"));
+  string fontpath = textLoader -> getString(string("IDS_TFP"));
   Button balance = Button(windowX, windowY, xPos, yPos,
     textLoader -> getString(string("IDS_Balance_Button_Text")) + to_string(startingBalance), textLoader, fontpath);
 
@@ -215,6 +290,9 @@ void PlayingScreenHeader::initBalanceButton(){
 
   //push this button into the vector of buttons to be drawn
   headerVariables.push_back(balance);
+
+  //push the initial value of the button
+  headerVariableValues.push_back(startingBalance);
 }
 
 /*
@@ -243,7 +321,7 @@ void PlayingScreenHeader::initLevelButton(){
   float yPos =  headerButtonPaddingY;
 
 
-  string fontpath = textLoader -> getString(string("IDS_FFP"));
+  string fontpath = textLoader -> getString(string("IDS_TFP"));
   Button level = Button(windowX, windowY, xPos, yPos,
     textLoader -> getString(string("IDS_Level_Button_Text")) + to_string(startingLevel), textLoader, fontpath);
 
@@ -285,6 +363,9 @@ void PlayingScreenHeader::initLevelButton(){
 
   //push this button into the vector of buttons to be drawn
   headerVariables.push_back(level);
+
+  //push the inital value
+  headerVariableValues.push_back(startingLevel);
 }
 
 /*
@@ -313,7 +394,7 @@ void PlayingScreenHeader::initWaveButton(){
   float yPos =  headerButtonPaddingY;
 
 
-  string fontpath = textLoader -> getString(string("IDS_FFP"));
+  string fontpath = textLoader -> getString(string("IDS_TFP"));
   Button wave = Button(windowX, windowY, xPos, yPos,
     textLoader -> getString(string("IDS_Wave_Button_Text")) + to_string(startingWave), textLoader, fontpath);
 
@@ -355,17 +436,176 @@ void PlayingScreenHeader::initWaveButton(){
 
   //push this button into the vector of buttons to be drawn
   headerVariables.push_back(wave);
+
+  //push the inital value
+  headerVariableValues.push_back(startingWave);
+}
+
+/*
+ * Recompute the size and strings for the buttons when
+ * one of their values has been updated
+ */
+void PlayingScreenHeader::resetButtons(){
+
+  resetHitpointButton();
+  resetBalanceButton();
+  resetLevelButton();
+  resetWaveButton();
+
+}
+
+/*
+ * Recompute the size and string of the hitpoints button
+ */
+void PlayingScreenHeader::resetHitpointButton(){
+
+    //padding between each button in header
+    int headerButtonPaddingX = textLoader -> getInteger(string("IDS_Playing_Screen_Header_Padding_X"));
+    int headerButtonPaddingY = textLoader -> getInteger(string("IDS_Playing_Screen_Header_Padding_Y"));
+
+    //x position is from the left plus padding
+    float xPos = headerButtonPaddingX;
+    //y position is from the top plus padding
+    float yPos = headerButtonPaddingY;
+
+    //set the string
+    headerVariables.at(0).setString(textLoader -> getString("IDS_Hitpoints_Button_Text") + to_string(headerVariableValues.at(0)));
+
+    //rescale the button and reset it
+    headerVariables.at(0).scaleButton(xPos, yPos);
+
+}
+
+/*
+ * Recompute the size and string of the balance button
+ */
+void PlayingScreenHeader::resetBalanceButton(){
+    //the bounding rectangle for the rectangle around the balance button
+    sf::RectangleShape offSetRect = headerVariables.at(0).getButtonRect();
+
+    //position
+    sf::Vector2f pos = offSetRect.getPosition();
+    //dimensions
+    sf::Vector2f dimensions = offSetRect.getSize();
+
+    //padding between each button in header
+    int headerButtonPaddingX = textLoader -> getInteger(string("IDS_Playing_Screen_Header_Padding_X"));
+    int headerButtonPaddingY = textLoader -> getInteger(string("IDS_Playing_Screen_Header_Padding_Y"));
+
+    //the xPosition is the length of the balance button plus some padding
+    float xPos = pos.x + dimensions.x + headerButtonPaddingX;
+    //the yPosition is the
+    float yPos =  headerButtonPaddingY;
+
+    //set the string
+    headerVariables.at(1).setString(textLoader -> getString("IDS_Balance_Button_Text") + to_string(headerVariableValues.at(1)));
+
+    //rescale the button and reset it
+    headerVariables.at(1).scaleButton(xPos, yPos);
+
+}
+
+/*
+ * Recompute the size and string of the level button
+ */
+void PlayingScreenHeader::resetLevelButton(){
+    //the bounding rectangle for the rectangle around the wave button
+    sf::RectangleShape offSetRect = headerVariables.at(1).getButtonRect();
+
+    //position
+    sf::Vector2f pos = offSetRect.getPosition();
+    //dimensions
+    sf::Vector2f dimensions = offSetRect.getSize();
+
+    //padding between each button in header
+    int headerButtonPaddingX = textLoader -> getInteger(string("IDS_Playing_Screen_Header_Padding_X"));
+    int headerButtonPaddingY = textLoader -> getInteger(string("IDS_Playing_Screen_Header_Padding_Y"));
+
+    //the xPosition is the length of the wave button plus some padding
+    float xPos = pos.x + dimensions.x + headerButtonPaddingX;
+    //the yPosition is the
+    float yPos =  headerButtonPaddingY;
+
+    //set the string
+    headerVariables.at(2).setString(textLoader -> getString("IDS_Level_Button_Text") + to_string(headerVariableValues.at(2)));
+
+    //rescale the button and reset it
+    headerVariables.at(2).scaleButton(xPos, yPos);
+
+}
+
+/*
+ * Recompute the size and string of the wave button
+ */
+void PlayingScreenHeader::resetWaveButton(){
+
+    //padding between each button in header
+    int headerButtonPaddingX = textLoader -> getInteger(string("IDS_Playing_Screen_Header_Padding_X"));
+    int headerButtonPaddingY = textLoader -> getInteger(string("IDS_Playing_Screen_Header_Padding_Y"));
+
+    //the bounding rectangle for the rectangle around the level button
+    sf::RectangleShape offSetRect = headerVariables.at(2).getButtonRect();
+
+    //position
+    sf::Vector2f pos = offSetRect.getPosition();
+    //dimensions
+    sf::Vector2f dimensions = offSetRect.getSize();
+
+    //the xPosition is the length of the level button plus some padding
+    float xPos = pos.x + dimensions.x + headerButtonPaddingX;
+    //the yPosition is the
+    float yPos =  headerButtonPaddingY;
+
+
+    //set the string
+    headerVariables.at(3).setString(textLoader -> getString("IDS_Wave_Button_Text") + to_string(headerVariableValues.at(3)));
+
+    //rescale the button and reset it
+    headerVariables.at(3).scaleButton(xPos, yPos);
+
 }
 
 /*
  * @return the maximum y dimension of any object in the header
  */
 float PlayingScreenHeader::getMaximumY(){
-  sf::RectangleShape rect = buyTower -> getButtonRect();
+
+  /*
+  //if the sell tower is visible it will be the lowest
+  if((this->sellTower)->isCurrentlyVisible()){
+    sf::RectangleShape rect = sellTower -> getButtonRect();
+    sf::Vector2f size = rect.getSize();
+    //position
+    sf::Vector2f pos = rect.getPosition();
+
+    int playingScreenHeaderPaddingY = textLoader->getInteger(string("IDS_Playing_Screen_Header_Padding_Y"));
+    return pos.y + size.y + playingScreenHeaderPaddingY;
+  }
+  //if the buy tower is visible it will be the lowest
+  else if((this->buyTower)->isCurrentlyVisible()){
+    sf::RectangleShape rect = buyTower -> getButtonRect();
+    sf::Vector2f size = rect.getSize();
+    //position
+    sf::Vector2f pos = rect.getPosition();
+
+    int playingScreenHeaderPaddingY = textLoader->getInteger(string("IDS_Playing_Screen_Header_Padding_Y"));
+    return pos.y + size.y + playingScreenHeaderPaddingY;
+  }
+  //otherwise it is one of the normal headers
+  else{
+    sf::RectangleShape rect = headerVariables.at(0).getButtonRect();
+    sf::Vector2f size = rect.getSize();
+
+    int playingScreenHeaderPaddingY = textLoader->getInteger(string("IDS_Playing_Screen_Header_Padding_Y"));
+    return size.y + playingScreenHeaderPaddingY;
+  }
+  */
+  //using a normal header for now since the resizing is looking weird
+  sf::RectangleShape rect = headerVariables.at(0).getButtonRect();
   sf::Vector2f size = rect.getSize();
 
-  int playingScreenHeaderPadding = textLoader->getInteger(string("IDS_Playing_Screen_Header_Padding"));
-  return size.y + playingScreenHeaderPadding;
+  int playingScreenHeaderPaddingY = textLoader->getInteger(string("IDS_Playing_Screen_Header_Padding_Y"));
+  return size.y + playingScreenHeaderPaddingY;
 }
 
 
@@ -392,6 +632,46 @@ void PlayingScreenHeader::registerDelegates(){
   EventType mousePressEventType = mousePressEvent.getEventType();
   //register the delegate and its type
   this -> eventManager -> registerDelegate(mousePressDelegate, textLoader -> getString(string("IDS_PSH_MP")),mousePressEventType);
+
+  //bind our delegate function for lose hitpoints
+  EventManager::EventDelegate lostHitpointsDelegate = std::bind(&PlayingScreenHeader::handleLoseHitpoints, this, _1);
+
+  //make an event and get its type
+  LoseHitpointsEvent loseHitpointsEvent = LoseHitpointsEvent();
+  EventType loseHitpointsEventType = loseHitpointsEvent.getEventType();
+  //register the delegate and its type
+  this -> eventManager -> registerDelegate(lostHitpointsDelegate, textLoader -> getString(string("IDS_PSH_LHP")),loseHitpointsEventType);
+
+
+  //bind our delegate function for balance change
+  EventManager::EventDelegate balanceChangeDelegate = std::bind(&PlayingScreenHeader::handleBalanceChange, this, _1);
+
+  //make an event and get its type
+  BalanceChangeEvent balanceChangeEvent = BalanceChangeEvent();
+  EventType balanceChangeEventType = balanceChangeEvent.getEventType();
+  //register the delegate and its type
+  this -> eventManager -> registerDelegate(balanceChangeDelegate, textLoader -> getString(string("IDS_PSH_BC")),balanceChangeEventType);
+
+
+  //bind our delegate function for level change
+  EventManager::EventDelegate levelChangeDelegate = std::bind(&PlayingScreenHeader::handleLevelChange, this, _1);
+
+  //make an event and get its type
+  LevelChangeEvent levelChangeEvent = LevelChangeEvent();
+  EventType levelChangeEventType = levelChangeEvent.getEventType();
+  //register the delegate and its type
+  this -> eventManager -> registerDelegate(levelChangeDelegate, textLoader -> getString(string("IDS_PSH_LC")),levelChangeEventType);
+
+
+  //bind our delegate function for wave change
+  EventManager::EventDelegate waveChangeDelegate = std::bind(&PlayingScreenHeader::handleWaveChange, this, _1);
+
+  //make an event and get its type
+  WaveChangeEvent waveChangeEvent = WaveChangeEvent();
+  EventType waveChangeEventType = waveChangeEvent.getEventType();
+  //register the delegate and its type
+  this -> eventManager -> registerDelegate(waveChangeDelegate, textLoader -> getString(string("IDS_PSH_WC")),waveChangeEventType);
+
 }
 
 /*
@@ -412,6 +692,29 @@ void PlayingScreenHeader::deregisterDelegates(){
   //register the delegate and its type
   this -> eventManager -> deregisterDelegate(textLoader -> getString(string("IDS_PSH_MP")),mousePressEventType);
 
+  //make an event and get its type
+  LoseHitpointsEvent loseHitpointsEvent = LoseHitpointsEvent();
+  EventType loseHitpointsEventType = loseHitpointsEvent.getEventType();
+  //deregister the delegate and its type
+  this -> eventManager -> deregisterDelegate( textLoader -> getString(string("IDS_PSH_LHP")),loseHitpointsEventType);
+
+  //make an event and get its type
+  BalanceChangeEvent balanceChangeEvent = BalanceChangeEvent();
+  EventType balanceChangeEventType = balanceChangeEvent.getEventType();
+  //deregister the delegate and its type
+  this -> eventManager -> deregisterDelegate( textLoader -> getString(string("IDS_PSH_BC")),balanceChangeEventType);
+
+  //make an event and get its type
+  LevelChangeEvent levelChangeEvent = LevelChangeEvent();
+  EventType levelChangeEventType = levelChangeEvent.getEventType();
+  //deregister the delegate and its type
+  this -> eventManager -> deregisterDelegate( textLoader -> getString(string("IDS_PSH_LC")),levelChangeEventType);
+
+  //make an event and get its type
+  WaveChangeEvent waveChangeEvent = WaveChangeEvent();
+  EventType waveChangeEventType = waveChangeEvent.getEventType();
+  //deregister the delegate and its type
+  this -> eventManager -> deregisterDelegate( textLoader -> getString(string("IDS_PSH_WC")),waveChangeEventType);
 }
 
 void PlayingScreenHeader::setSelectedTile(int row, int col){
@@ -486,11 +789,23 @@ void PlayingScreenHeader::handleMousePress(const EventInterface& event){
   //otherwise it will not register the click
   bool buyTowerClicked = (this->buyTower) -> isSelected(xPos + getXOffSet(),yPos + getYOffSet());
   if(buyTowerClicked){
-    shared_ptr<EventInterface> buyTowerState = make_shared<StateChangeEvent>(State::BuyTower, row, col, nowInNano);
+    shared_ptr<EventInterface> buyTowerState = make_shared<StateChangeEvent>(State::BuyTower, rowSelected, colSelected, true, nowInNano);
 
     this -> eventManager -> queueEvent(buyTowerState);
     return;
   }
+
+  //check if SellTower button is being clicked
+  //modify the xposition to reflect if there is an offset
+  //otherwise it will not register the click
+  bool sellTowerClicked = (this->sellTower) -> isSelected(xPos + getXOffSet(),yPos + getYOffSet());
+  if(sellTowerClicked){
+    shared_ptr<EventInterface> buyTowerState = make_shared<StateChangeEvent>(State::BuyTower, rowSelected, colSelected, false, nowInNano);
+
+    this -> eventManager -> queueEvent(buyTowerState);
+    return;
+  }
+
 
   //only check if we are clickinga tile when the cursor is within the game
   //area, subtracting the offset will render some clicks outside
@@ -512,22 +827,143 @@ void PlayingScreenHeader::handleMousePress(const EventInterface& event){
     else if(gameLogic->isExit(row,col) || gameLogic -> isPath(row,col)) {
       if((this->buyTower)->isCurrentlyVisible()){
           (this->buyTower)->flipVisibility();
+          recaculateHeader = true;
       }
       rowSelected = row;
       colSelected = col;
       return;
     }
 
+    //if the selected row and column is not the same as the old ones
+    //and/or the selected row col is not an existing tower
+    //then we hide the sell tower button
+    if(row==rowSelected && colSelected == col &&  gameLogic ->isTower(row,col)){
+      (this->sellTower)->flipVisibility();
+      recaculateHeader = true;
+    }
+    else if(!(gameLogic->isTower(row,col))){
+      if((this->sellTower)->isCurrentlyVisible()){
+        (this->sellTower)->flipVisibility();
+        recaculateHeader = true;
+      }
+    }
+
     //if we have selected the same row and col as the last click then we
     //flip the visibility of the button
     if(row == rowSelected && colSelected == col){
       (this->buyTower)->flipVisibility();
+      recaculateHeader = true;
     }
     else{
+      if((this->buyTower)->isCurrentlyVisible()){
+        (this->buyTower)->flipVisibility();
+        recaculateHeader = true;
+      }
       rowSelected = row;
       colSelected = col;
     }
   }
+}
+
+/*
+ * Handle any lost hitpoints from enemy units reaching exit
+ * @param event: event of the hitpoint loss
+ */
+void PlayingScreenHeader::handleLoseHitpoints(const EventInterface& event){
+  /*
+   * cast the EventInterface reference to a CONST pointer to the
+   * LoseHitpointsEvent type which allows us to access variables and methods
+   * specific to LoseHitpointsEvent
+   */
+  const LoseHitpointsEvent* lhpEvent = static_cast<const LoseHitpointsEvent*>(&event);
+  /*
+   * cast the "data" (a EventDataInterface) to a LoseHitpointsEventData type
+   * the .get() is because data is a unique_ptr and we need to grab the
+   * raw pointer inside of it for this
+   */
+  LoseHitpointsEventData* lhpEventData = static_cast<LoseHitpointsEventData*>((lhpEvent -> data).get());
+  //get the hitpoints lost from the data
+  int lostHitpoints = lhpEventData -> lostHitpoints;
+
+  headerVariableValues.at(0) -= lostHitpoints;
+
+  recaculateHeader=true;
+}
+
+/*
+ * Handle any change in the balance
+ * @param event: event of the balance change
+ */
+void PlayingScreenHeader::handleBalanceChange(const EventInterface& event){
+  /*
+   * cast the EventInterface reference to a CONST pointer to the
+   * BalanceChangeEvent type which allows us to access variables and methods
+   * specific to BalanceChangeEvent
+   */
+  const BalanceChangeEvent* bcEvent = static_cast<const BalanceChangeEvent*>(&event);
+  /*
+   * cast the "data" (a EventDataInterface) to a BalanceChangeEventData type
+   * the .get() is because data is a unique_ptr and we need to grab the
+   * raw pointer inside of it for this
+   */
+  BalanceChangeEventData* bcEventData = static_cast<BalanceChangeEventData*>((bcEvent -> data).get());
+  //get the balance change
+  int balanceChange = bcEventData -> balanceChange;
+
+  headerVariableValues.at(1) += balanceChange;
+
+  recaculateHeader=true;
+}
+
+
+/*
+ * Handle any level change
+ * @param event: event of the level change
+ */
+void PlayingScreenHeader::handleLevelChange(const EventInterface& event){
+  /*
+   * cast the EventInterface reference to a CONST pointer to the
+   * LevelChangeEvent type which allows us to access variables and methods
+   * specific to LevelChangeEvent
+   */
+  const LevelChangeEvent* lcEvent = static_cast<const LevelChangeEvent*>(&event);
+  /*
+   * cast the "data" (a EventDataInterface) to a LevelChangeEventData type
+   * the .get() is because data is a unique_ptr and we need to grab the
+   * raw pointer inside of it for this
+   */
+  LevelChangeEventData* lcEventData = static_cast<LevelChangeEventData*>((lcEvent -> data).get());
+  //get the new level value
+  int newLevel = lcEventData -> level;
+
+  headerVariableValues.at(2) = newLevel;
+
+  recaculateHeader=true;
+}
+
+/*
+ * Handle wave change
+ * @param event: event of next wave
+ */
+void PlayingScreenHeader::handleWaveChange(const EventInterface& event){
+  /*
+   * cast the EventInterface reference to a CONST pointer to the
+   * WaveChangeEvent type which allows us to access variables and methods
+   * specific to WaveChangeEvent
+   */
+  const WaveChangeEvent* wcEvent = static_cast<const WaveChangeEvent*>(&event);
+  /*
+   * cast the "data" (a EventDataInterface) to a WaveChangeEventData type
+   * the .get() is because data is a unique_ptr and we need to grab the
+   * raw pointer inside of it for this
+   */
+  WaveChangeEventData* wcEventData = static_cast<WaveChangeEventData*>((wcEvent -> data).get());
+  //get the new wave
+  int newWave = wcEventData -> waveNum;
+
+  headerVariableValues.at(3) = newWave;
+
+  recaculateHeader=true;
 }
 
 /*
@@ -579,8 +1015,16 @@ float PlayingScreenHeader::getYOffSet(){
 }
 
 void PlayingScreenHeader::draw(sf::RenderWindow &window){
-  //handle the buy tower button separately
+  //if something has changed we need to resize the header
+  if(recaculateHeader){
+    ySize = getMaximumY();
+    resetButtons();
+    recaculateHeader = false;
+  }
+
+  //handle the buy tower and sell tower buttons separately
   drawBuyTowerButton(window);
+  drawSellTowerButton(window);
   //draw the rest of the header if it is visible
   if(visible){
     drawHeaderButtons(window);
@@ -600,7 +1044,7 @@ void PlayingScreenHeader::drawBuyTowerButton(sf::RenderWindow& window){
   }
 
   //used to make the font local
-  string mainFontPath = textLoader -> getString(string("IDS_FFP"));
+  string mainFontPath = textLoader -> getString(string("IDS_TFP"));
 
   if(!mainFont.loadFromFile(mainFontPath)){
     cout << "No font!" << endl;
@@ -626,6 +1070,22 @@ void PlayingScreenHeader::drawSellTowerButton(sf::RenderWindow& window){
   if(!sellTower -> isCurrentlyVisible()){
     return;
   }
+
+  //used to make the font local
+  string mainFontPath = textLoader -> getString(string("IDS_TFP"));
+
+  if(!mainFont.loadFromFile(mainFontPath)){
+    cout << "No font!" << endl;
+  }
+  else{
+  //  cout << "loaded font!" << endl;
+  }
+
+  text = sellTower -> getButtonText();
+  text.setFont(mainFont);
+
+  window.draw(sellTower -> getButtonRect());
+  window.draw(text);
 }
 
 /*